DOPE-226: renamed variables, renamed test package
