+++ core/src/main/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/arrayfunction/ArrayAppendExpression.kt
@@ -7,11 +7,11 @@ import ch.ergon.dope.validtype.ValidType
 class ArrayAppendExpression<T : ValidType>(
     array: TypeExpression<ArrayType<T>>,
     value: TypeExpression<T>,
-    vararg values: TypeExpression<T>,
-) : ArrayFunctionExpression<T>("ARRAY_APPEND", array, value, *values)
+    vararg additionalValues: TypeExpression<T>,
+) : ArrayFunctionExpression<T>("ARRAY_APPEND", array, value, *additionalValues)
 
 fun <T : ValidType> arrayAppend(
     array: TypeExpression<ArrayType<T>>,
     value: TypeExpression<T>,
-    vararg values: TypeExpression<T>,
-) = ArrayAppendExpression(array, value, *values)
+    vararg additionalValues: TypeExpression<T>,
+) = ArrayAppendExpression(array, value, *additionalValues)
