+++ core/src/main/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/arrayfunction/ArrayRemoveExpression.kt
@@ -7,11 +7,11 @@ import ch.ergon.dope.validtype.ValidType
 class ArrayRemoveExpression<T : ValidType>(
     array: TypeExpression<ArrayType<T>>,
     value: TypeExpression<T>,
-    vararg values: TypeExpression<T>,
-) : ArrayFunctionExpression<T>("ARRAY_REMOVE", array, value, *values)
+    vararg additionalValues: TypeExpression<T>,
+) : ArrayFunctionExpression<T>("ARRAY_REMOVE", array, value, *additionalValues)
 
 fun <T : ValidType> arrayRemove(
     array: TypeExpression<ArrayType<T>>,
     value: TypeExpression<T>,
-    vararg values: TypeExpression<T>,
-) = ArrayRemoveExpression(array, value, *values)
+    vararg additionalValues: TypeExpression<T>,
+) = ArrayRemoveExpression(array, value, *additionalValues)
