+++ core/src/test/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/arrayfunction/ArrayAppendExpressionTest.kt
@@ -1,28 +1,27 @@
-package ch.ergon.dope.function.array
+package ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction
 
 import ch.ergon.dope.DopeQuery
 import ch.ergon.dope.helper.someNumberArrayField
+import ch.ergon.dope.helper.someNumberField
 import ch.ergon.dope.resolvable.expression.unaliased.type.ParameterManager
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.ArrayIntersectExpression
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.arrayIntersect
 import ch.ergon.dope.resolvable.expression.unaliased.type.asParameter
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class ArrayIntersectExpressionTest {
+class ArrayAppendExpressionTest {
     @BeforeEach
     fun reset() {
         ParameterManager.resetCounter()
     }
 
     @Test
-    fun `should support ARRAY_INTERSECT`() {
+    fun `should support ARRAY_APPEND`() {
         val expected = DopeQuery(
-            "ARRAY_INTERSECT(`numberArrayField`, `numberArrayField`)",
+            "ARRAY_APPEND(`numberArrayField`, `numberField`)",
             emptyMap(),
         )
-        val underTest = ArrayIntersectExpression(someNumberArrayField(), someNumberArrayField())
+        val underTest = ArrayAppendExpression(someNumberArrayField(), someNumberField())
 
         val actual = underTest.toDopeQuery()
 
@@ -30,13 +29,13 @@ class ArrayIntersectExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_INTERSECT with parameter`() {
+    fun `should support ARRAY_APPEND with parameter`() {
         val parameterValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            "ARRAY_INTERSECT($1, `numberArrayField`)",
+            "ARRAY_APPEND($1, `numberField`)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayIntersectExpression(parameterValue.asParameter(), someNumberArrayField())
+        val underTest = ArrayAppendExpression(parameterValue.asParameter(), someNumberField())
 
         val actual = underTest.toDopeQuery()
 
@@ -44,13 +43,13 @@ class ArrayIntersectExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_INTERSECT with parameter as value`() {
-        val parameterValue = listOf(1, 2, 3)
+    fun `should support ARRAY_APPEND with parameter as value`() {
+        val parameterValue = 1
         val expected = DopeQuery(
-            "ARRAY_INTERSECT(`numberArrayField`, $1)",
+            "ARRAY_APPEND(`numberArrayField`, $1)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayIntersectExpression(someNumberArrayField(), parameterValue.asParameter())
+        val underTest = ArrayAppendExpression(someNumberArrayField(), parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -58,14 +57,14 @@ class ArrayIntersectExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_INTERSECT with all parameters`() {
+    fun `should support ARRAY_APPEND with all parameters`() {
         val parameterValueCollection = listOf(1, 2, 3)
-        val parameterValue = listOf(4, 5, 6)
+        val parameterValue = 1
         val expected = DopeQuery(
-            "ARRAY_INTERSECT($1, $2)",
+            "ARRAY_APPEND($1, $2)",
             mapOf("$1" to parameterValueCollection, "$2" to parameterValue),
         )
-        val underTest = ArrayIntersectExpression(parameterValueCollection.asParameter(), parameterValue.asParameter())
+        val underTest = ArrayAppendExpression(parameterValueCollection.asParameter(), parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -73,12 +72,12 @@ class ArrayIntersectExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_INTERSECT extension`() {
+    fun `should support ARRAY_APPEND extension`() {
         val array = someNumberArrayField()
-        val value = someNumberArrayField()
-        val expected = ArrayIntersectExpression(array, value)
+        val value = someNumberField()
+        val expected = ArrayAppendExpression(array, value)
 
-        val actual = arrayIntersect(array, value)
+        val actual = arrayAppend(array, value)
 
         assertEquals(expected.toDopeQuery(), actual.toDopeQuery())
     }
