+++ core/src/test/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/arrayfunction/ArrayAverageExpressionTest.kt
@@ -1,29 +1,26 @@
-package ch.ergon.dope.function.array
+package ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction
 
 import ch.ergon.dope.DopeQuery
 import ch.ergon.dope.helper.someNumberArrayField
 import ch.ergon.dope.resolvable.expression.unaliased.type.ParameterManager
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.ArrayIfNullExpression
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.arrayIfNull
 import ch.ergon.dope.resolvable.expression.unaliased.type.asParameter
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class ArrayIfNullExpressionTest {
+class ArrayAverageExpressionTest {
     @BeforeEach
     fun reset() {
         ParameterManager.resetCounter()
     }
 
     @Test
-    fun `should support ARRAY_IFNULL`() {
-        val array = someNumberArrayField()
+    fun `should support ARRAY_AVG`() {
         val expected = DopeQuery(
-            "ARRAY_IFNULL(`numberArrayField`)",
+            "ARRAY_AVG(`numberArrayField`)",
             emptyMap(),
         )
-        val underTest = ArrayIfNullExpression(array)
+        val underTest = ArrayAverageExpression(someNumberArrayField())
 
         val actual = underTest.toDopeQuery()
 
@@ -31,13 +28,13 @@ class ArrayIfNullExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_IFNULL with parameter`() {
+    fun `should support ARRAY_AVG with parameter`() {
         val parameterValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            "ARRAY_IFNULL($1)",
+            "ARRAY_AVG($1)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayIfNullExpression(parameterValue.asParameter())
+        val underTest = ArrayAverageExpression(parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -45,11 +42,11 @@ class ArrayIfNullExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_IFNULL extension`() {
+    fun `should support ARRAY_AVG extension`() {
         val array = someNumberArrayField()
-        val expected = ArrayIfNullExpression(array)
+        val expected = ArrayAverageExpression(array)
 
-        val actual = arrayIfNull(array)
+        val actual = arrayAverage(array)
 
         assertEquals(expected.toDopeQuery(), actual.toDopeQuery())
     }
