+++ core/src/test/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/arrayfunction/ArrayConcatExpressionTest.kt
@@ -1,29 +1,26 @@
-package ch.ergon.dope.function.array
+package ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction
 
 import ch.ergon.dope.DopeQuery
 import ch.ergon.dope.helper.someNumberArrayField
-import ch.ergon.dope.helper.someNumberField
 import ch.ergon.dope.resolvable.expression.unaliased.type.ParameterManager
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.ArrayPutExpression
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.arrayPut
 import ch.ergon.dope.resolvable.expression.unaliased.type.asParameter
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class ArrayPutExpressionTest {
+class ArrayConcatExpressionTest {
     @BeforeEach
     fun reset() {
         ParameterManager.resetCounter()
     }
 
     @Test
-    fun `should support ARRAY_PUT`() {
+    fun `should support ARRAY_CONCAT`() {
         val expected = DopeQuery(
-            "ARRAY_PUT(`numberArrayField`, `numberField`)",
+            "ARRAY_CONCAT(`numberArrayField`, `numberArrayField`)",
             emptyMap(),
         )
-        val underTest = ArrayPutExpression(someNumberArrayField(), someNumberField())
+        val underTest = ArrayConcatExpression(someNumberArrayField(), someNumberArrayField())
 
         val actual = underTest.toDopeQuery()
 
@@ -31,13 +28,13 @@ class ArrayPutExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_PUT with parameter`() {
+    fun `should support ARRAY_CONCAT with parameter`() {
         val parameterValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            "ARRAY_PUT($1, `numberField`)",
+            "ARRAY_CONCAT($1, `numberArrayField`)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayPutExpression(parameterValue.asParameter(), someNumberField())
+        val underTest = ArrayConcatExpression(parameterValue.asParameter(), someNumberArrayField())
 
         val actual = underTest.toDopeQuery()
 
@@ -45,13 +42,13 @@ class ArrayPutExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_PUT with parameter as value`() {
-        val parameterValue = 1
+    fun `should support ARRAY_CONCAT with parameter as value`() {
+        val parameterValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            "ARRAY_PUT(`numberArrayField`, $1)",
+            "ARRAY_CONCAT(`numberArrayField`, $1)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayPutExpression(someNumberArrayField(), parameterValue.asParameter())
+        val underTest = ArrayConcatExpression(someNumberArrayField(), parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -59,14 +56,14 @@ class ArrayPutExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_PUT with all parameters`() {
+    fun `should support ARRAY_CONCAT with all parameters`() {
         val parameterValueCollection = listOf(1, 2, 3)
-        val parameterValue = 1
+        val parameterValue = listOf(4, 5, 6)
         val expected = DopeQuery(
-            "ARRAY_PUT($1, $2)",
+            "ARRAY_CONCAT($1, $2)",
             mapOf("$1" to parameterValueCollection, "$2" to parameterValue),
         )
-        val underTest = ArrayPutExpression(parameterValueCollection.asParameter(), parameterValue.asParameter())
+        val underTest = ArrayConcatExpression(parameterValueCollection.asParameter(), parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -74,12 +71,12 @@ class ArrayPutExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_PUT extension`() {
+    fun `should support ARRAY_CONCAT extension`() {
         val array = someNumberArrayField()
-        val value = someNumberField()
-        val expected = ArrayPutExpression(array, value)
+        val value = someNumberArrayField()
+        val expected = ArrayConcatExpression(array, value)
 
-        val actual = arrayPut(array, value)
+        val actual = arrayConcat(array, value)
 
         assertEquals(expected.toDopeQuery(), actual.toDopeQuery())
     }
