+++ core/src/test/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/arrayfunction/ArrayMaxExpressionTest.kt
@@ -1,28 +1,26 @@
-package ch.ergon.dope.function.array
+package ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction
 
 import ch.ergon.dope.DopeQuery
 import ch.ergon.dope.helper.someNumberArrayField
 import ch.ergon.dope.resolvable.expression.unaliased.type.ParameterManager
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.ArrayMinExpression
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.arrayMin
 import ch.ergon.dope.resolvable.expression.unaliased.type.asParameter
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class ArrayMinExpressionTest {
+class ArrayMaxExpressionTest {
     @BeforeEach
     fun reset() {
         ParameterManager.resetCounter()
     }
 
     @Test
-    fun `should support ARRAY_MIN`() {
+    fun `should support ARRAY_MAX`() {
         val expected = DopeQuery(
-            "ARRAY_MIN(`numberArrayField`)",
+            "ARRAY_MAX(`numberArrayField`)",
             emptyMap(),
         )
-        val underTest = ArrayMinExpression(someNumberArrayField())
+        val underTest = ArrayMaxExpression(someNumberArrayField())
 
         val actual = underTest.toDopeQuery()
 
@@ -30,13 +28,13 @@ class ArrayMinExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_MIN with parameter`() {
+    fun `should support ARRAY_MAX with parameter`() {
         val parameterValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            "ARRAY_MIN($1)",
+            "ARRAY_MAX($1)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayMinExpression(parameterValue.asParameter())
+        val underTest = ArrayMaxExpression(parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -44,11 +42,11 @@ class ArrayMinExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_MIN extension`() {
+    fun `should support ARRAY_MAX extension`() {
         val array = someNumberArrayField()
-        val expected = ArrayMinExpression(array)
+        val expected = ArrayMaxExpression(array)
 
-        val actual = arrayMin(array)
+        val actual = arrayMax(array)
 
         assertEquals(expected.toDopeQuery(), actual.toDopeQuery())
     }
