+++ core/src/test/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/arrayfunction/ArrayMinExpressionTest.kt
@@ -1,28 +1,26 @@
-package ch.ergon.dope.function.array
+package ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction
 
 import ch.ergon.dope.DopeQuery
 import ch.ergon.dope.helper.someNumberArrayField
 import ch.ergon.dope.resolvable.expression.unaliased.type.ParameterManager
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.ArrayDistinctExpression
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.arrayDistinct
 import ch.ergon.dope.resolvable.expression.unaliased.type.asParameter
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class ArrayDistinctExpressionTest {
+class ArrayMinExpressionTest {
     @BeforeEach
     fun reset() {
         ParameterManager.resetCounter()
     }
 
     @Test
-    fun `should support ARRAY_DISTINCT`() {
+    fun `should support ARRAY_MIN`() {
         val expected = DopeQuery(
-            "ARRAY_DISTINCT(`numberArrayField`)",
+            "ARRAY_MIN(`numberArrayField`)",
             emptyMap(),
         )
-        val underTest = ArrayDistinctExpression(someNumberArrayField())
+        val underTest = ArrayMinExpression(someNumberArrayField())
 
         val actual = underTest.toDopeQuery()
 
@@ -30,13 +28,13 @@ class ArrayDistinctExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_DISTINCT with parameter`() {
+    fun `should support ARRAY_MIN with parameter`() {
         val parameterValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            "ARRAY_DISTINCT($1)",
+            "ARRAY_MIN($1)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayDistinctExpression(parameterValue.asParameter())
+        val underTest = ArrayMinExpression(parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -44,11 +42,11 @@ class ArrayDistinctExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_DISTINCT extension`() {
+    fun `should support ARRAY_MIN extension`() {
         val array = someNumberArrayField()
-        val expected = ArrayDistinctExpression(array)
+        val expected = ArrayMinExpression(array)
 
-        val actual = arrayDistinct(array)
+        val actual = arrayMin(array)
 
         assertEquals(expected.toDopeQuery(), actual.toDopeQuery())
     }
