+++ core/src/test/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/arrayfunction/ArrayMoveExpressionTest.kt
@@ -1,10 +1,8 @@
-package ch.ergon.dope.function.array
+package ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction
 
 import ch.ergon.dope.DopeQuery
 import ch.ergon.dope.helper.someNumberArrayField
 import ch.ergon.dope.resolvable.expression.unaliased.type.ParameterManager
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.ArrayMoveExpression
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.arrayMove
 import ch.ergon.dope.resolvable.expression.unaliased.type.asParameter
 import ch.ergon.dope.resolvable.expression.unaliased.type.toDopeType
 import org.junit.jupiter.api.Assertions.assertEquals
