+++ core/src/test/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/arrayfunction/ArrayPositionExpressionTest.kt
@@ -1,29 +1,28 @@
-package ch.ergon.dope.function.array
+package ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction
 
 import ch.ergon.dope.DopeQuery
 import ch.ergon.dope.helper.someNumberArrayField
 import ch.ergon.dope.helper.someNumberField
 import ch.ergon.dope.resolvable.expression.unaliased.type.ParameterManager
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.ArrayRemoveExpression
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.arrayRemove
 import ch.ergon.dope.resolvable.expression.unaliased.type.asParameter
+import ch.ergon.dope.resolvable.expression.unaliased.type.toDopeType
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class ArrayRemoveExpressionTest {
+class ArrayPositionExpressionTest {
     @BeforeEach
     fun reset() {
         ParameterManager.resetCounter()
     }
 
     @Test
-    fun `should support ARRAY_REMOVE`() {
+    fun `should support ARRAY_POSITION`() {
         val expected = DopeQuery(
-            "ARRAY_REMOVE(`numberArrayField`, `numberField`)",
+            "ARRAY_POSITION(`numberArrayField`, `numberField`)",
             emptyMap(),
         )
-        val underTest = ArrayRemoveExpression(someNumberArrayField(), someNumberField())
+        val underTest = ArrayPositionExpression(someNumberArrayField(), someNumberField())
 
         val actual = underTest.toDopeQuery()
 
@@ -31,13 +30,13 @@ class ArrayRemoveExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_REMOVE with parameter`() {
+    fun `should support ARRAY_POSITION with parameter`() {
         val parameterValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            "ARRAY_REMOVE($1, `numberField`)",
+            "ARRAY_POSITION($1, `numberField`)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayRemoveExpression(parameterValue.asParameter(), someNumberField())
+        val underTest = ArrayPositionExpression(parameterValue.asParameter(), someNumberField())
 
         val actual = underTest.toDopeQuery()
 
@@ -45,13 +44,13 @@ class ArrayRemoveExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_REMOVE with parameter as value`() {
+    fun `should support ARRAY_POSITION with parameter as value`() {
         val parameterValue = 1
         val expected = DopeQuery(
-            "ARRAY_REMOVE(`numberArrayField`, $1)",
+            "ARRAY_POSITION(`numberArrayField`, $1)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayRemoveExpression(someNumberArrayField(), parameterValue.asParameter())
+        val underTest = ArrayPositionExpression(someNumberArrayField(), parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -59,14 +58,14 @@ class ArrayRemoveExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_REMOVE with all parameters`() {
+    fun `should support ARRAY_POSITION with all parameters`() {
         val parameterValueCollection = listOf(1, 2, 3)
         val parameterValue = 1
         val expected = DopeQuery(
-            "ARRAY_REMOVE($1, $2)",
+            "ARRAY_POSITION($1, $2)",
             mapOf("$1" to parameterValueCollection, "$2" to parameterValue),
         )
-        val underTest = ArrayRemoveExpression(parameterValueCollection.asParameter(), parameterValue.asParameter())
+        val underTest = ArrayPositionExpression(parameterValueCollection.asParameter(), parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -74,12 +73,23 @@ class ArrayRemoveExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_REMOVE extension`() {
+    fun `should support ARRAY_POSITION extension`() {
         val array = someNumberArrayField()
         val value = someNumberField()
-        val expected = ArrayRemoveExpression(array, value)
+        val expected = ArrayPositionExpression(array, value)
 
-        val actual = arrayRemove(array, value)
+        val actual = arrayPosition(array, value)
+
+        assertEquals(expected.toDopeQuery(), actual.toDopeQuery())
+    }
+
+    @Test
+    fun `should support ARRAY_POSITION extension number`() {
+        val array = someNumberArrayField()
+        val value = 1
+        val expected = ArrayPositionExpression(array, value.toDopeType())
+
+        val actual = arrayPosition(array, value)
 
         assertEquals(expected.toDopeQuery(), actual.toDopeQuery())
     }
