+++ core/src/test/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/arrayfunction/ArrayRemoveExpressionTest.kt
@@ -1,29 +1,27 @@
-package ch.ergon.dope.function.array
+package ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction
 
 import ch.ergon.dope.DopeQuery
 import ch.ergon.dope.helper.someNumberArrayField
 import ch.ergon.dope.helper.someNumberField
 import ch.ergon.dope.resolvable.expression.unaliased.type.ParameterManager
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.ArrayAppendExpression
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.arrayAppend
 import ch.ergon.dope.resolvable.expression.unaliased.type.asParameter
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class ArrayAppendExpressionTest {
+class ArrayRemoveExpressionTest {
     @BeforeEach
     fun reset() {
         ParameterManager.resetCounter()
     }
 
     @Test
-    fun `should support ARRAY_APPEND`() {
+    fun `should support ARRAY_REMOVE`() {
         val expected = DopeQuery(
-            "ARRAY_APPEND(`numberArrayField`, `numberField`)",
+            "ARRAY_REMOVE(`numberArrayField`, `numberField`)",
             emptyMap(),
         )
-        val underTest = ArrayAppendExpression(someNumberArrayField(), someNumberField())
+        val underTest = ArrayRemoveExpression(someNumberArrayField(), someNumberField())
 
         val actual = underTest.toDopeQuery()
 
@@ -31,13 +29,13 @@ class ArrayAppendExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_APPEND with parameter`() {
+    fun `should support ARRAY_REMOVE with parameter`() {
         val parameterValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            "ARRAY_APPEND($1, `numberField`)",
+            "ARRAY_REMOVE($1, `numberField`)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayAppendExpression(parameterValue.asParameter(), someNumberField())
+        val underTest = ArrayRemoveExpression(parameterValue.asParameter(), someNumberField())
 
         val actual = underTest.toDopeQuery()
 
@@ -45,13 +43,13 @@ class ArrayAppendExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_APPEND with parameter as value`() {
+    fun `should support ARRAY_REMOVE with parameter as value`() {
         val parameterValue = 1
         val expected = DopeQuery(
-            "ARRAY_APPEND(`numberArrayField`, $1)",
+            "ARRAY_REMOVE(`numberArrayField`, $1)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayAppendExpression(someNumberArrayField(), parameterValue.asParameter())
+        val underTest = ArrayRemoveExpression(someNumberArrayField(), parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -59,14 +57,14 @@ class ArrayAppendExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_APPEND with all parameters`() {
+    fun `should support ARRAY_REMOVE with all parameters`() {
         val parameterValueCollection = listOf(1, 2, 3)
         val parameterValue = 1
         val expected = DopeQuery(
-            "ARRAY_APPEND($1, $2)",
+            "ARRAY_REMOVE($1, $2)",
             mapOf("$1" to parameterValueCollection, "$2" to parameterValue),
         )
-        val underTest = ArrayAppendExpression(parameterValueCollection.asParameter(), parameterValue.asParameter())
+        val underTest = ArrayRemoveExpression(parameterValueCollection.asParameter(), parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -74,12 +72,12 @@ class ArrayAppendExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_APPEND extension`() {
+    fun `should support ARRAY_REMOVE extension`() {
         val array = someNumberArrayField()
         val value = someNumberField()
-        val expected = ArrayAppendExpression(array, value)
+        val expected = ArrayRemoveExpression(array, value)
 
-        val actual = arrayAppend(array, value)
+        val actual = arrayRemove(array, value)
 
         assertEquals(expected.toDopeQuery(), actual.toDopeQuery())
     }
