+++ core/src/test/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/arrayfunction/ArrayReverseExpressionTest.kt
@@ -1,28 +1,26 @@
-package ch.ergon.dope.function.array
+package ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction
 
 import ch.ergon.dope.DopeQuery
 import ch.ergon.dope.helper.someNumberArrayField
 import ch.ergon.dope.resolvable.expression.unaliased.type.ParameterManager
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.ArrayLengthExpression
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.arrayLength
 import ch.ergon.dope.resolvable.expression.unaliased.type.asParameter
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class ArrayLengthExpressionTest {
+class ArrayReverseExpressionTest {
     @BeforeEach
     fun reset() {
         ParameterManager.resetCounter()
     }
 
     @Test
-    fun `should support ARRAY_LENGTH`() {
+    fun `should support ARRAY_REVERSE`() {
         val expected = DopeQuery(
-            "ARRAY_LENGTH(`numberArrayField`)",
+            "ARRAY_REVERSE(`numberArrayField`)",
             emptyMap(),
         )
-        val underTest = ArrayLengthExpression(someNumberArrayField())
+        val underTest = ArrayReverseExpression(someNumberArrayField())
 
         val actual = underTest.toDopeQuery()
 
@@ -30,13 +28,13 @@ class ArrayLengthExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_LENGTH with parameter`() {
+    fun `should support ARRAY_REVERSE with parameter`() {
         val parameterValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            "ARRAY_LENGTH($1)",
+            "ARRAY_REVERSE($1)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayLengthExpression(parameterValue.asParameter())
+        val underTest = ArrayReverseExpression(parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -44,11 +42,11 @@ class ArrayLengthExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_LENGTH extension`() {
+    fun `should support ARRAY_REVERSE extension`() {
         val array = someNumberArrayField()
-        val expected = ArrayLengthExpression(array)
+        val expected = ArrayReverseExpression(array)
 
-        val actual = arrayLength(array)
+        val actual = arrayReverse(array)
 
         assertEquals(expected.toDopeQuery(), actual.toDopeQuery())
     }
