+++ core/src/test/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/arrayfunction/ArraySumExpressionTest.kt
@@ -1,28 +1,26 @@
-package ch.ergon.dope.function.array
+package ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction
 
 import ch.ergon.dope.DopeQuery
 import ch.ergon.dope.helper.someNumberArrayField
 import ch.ergon.dope.resolvable.expression.unaliased.type.ParameterManager
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.ArrayAverageExpression
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.arrayAverage
 import ch.ergon.dope.resolvable.expression.unaliased.type.asParameter
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class ArrayAverageExpressionTest {
+class ArraySumExpressionTest {
     @BeforeEach
     fun reset() {
         ParameterManager.resetCounter()
     }
 
     @Test
-    fun `should support ARRAY_AVG`() {
+    fun `should support ARRAY_SUM`() {
         val expected = DopeQuery(
-            "ARRAY_AVG(`numberArrayField`)",
+            "ARRAY_SUM(`numberArrayField`)",
             emptyMap(),
         )
-        val underTest = ArrayAverageExpression(someNumberArrayField())
+        val underTest = ArraySumExpression(someNumberArrayField())
 
         val actual = underTest.toDopeQuery()
 
@@ -30,13 +28,13 @@ class ArrayAverageExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_AVG with parameter`() {
+    fun `should support ARRAY_SUM with parameter`() {
         val parameterValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            "ARRAY_AVG($1)",
+            "ARRAY_SUM($1)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayAverageExpression(parameterValue.asParameter())
+        val underTest = ArraySumExpression(parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -44,11 +42,11 @@ class ArrayAverageExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_AVG extension`() {
+    fun `should support ARRAY_SUM extension`() {
         val array = someNumberArrayField()
-        val expected = ArrayAverageExpression(array)
+        val expected = ArraySumExpression(array)
 
-        val actual = arrayAverage(array)
+        val actual = arraySum(array)
 
         assertEquals(expected.toDopeQuery(), actual.toDopeQuery())
     }
