+++ core/src/test/kotlin/ch/ergon/dope/resolvable/expression/unaliased/type/arrayfunction/ArrayUnionExpressionTest.kt
@@ -1,28 +1,26 @@
-package ch.ergon.dope.function.array
+package ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction
 
 import ch.ergon.dope.DopeQuery
 import ch.ergon.dope.helper.someNumberArrayField
 import ch.ergon.dope.resolvable.expression.unaliased.type.ParameterManager
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.ArrayExceptExpression
-import ch.ergon.dope.resolvable.expression.unaliased.type.arrayfunction.arrayExcept
 import ch.ergon.dope.resolvable.expression.unaliased.type.asParameter
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class ArrayExceptExpressionTest {
+class ArrayUnionExpressionTest {
     @BeforeEach
     fun reset() {
         ParameterManager.resetCounter()
     }
 
     @Test
-    fun `should support ARRAY_EXCEPT`() {
+    fun `should support ARRAY_UNION`() {
         val expected = DopeQuery(
-            "ARRAY_EXCEPT(`numberArrayField`, `numberArrayField`)",
+            "ARRAY_UNION(`numberArrayField`, `numberArrayField`)",
             emptyMap(),
         )
-        val underTest = ArrayExceptExpression(someNumberArrayField(), someNumberArrayField())
+        val underTest = ArrayUnionExpression(someNumberArrayField(), someNumberArrayField())
 
         val actual = underTest.toDopeQuery()
 
@@ -30,13 +28,13 @@ class ArrayExceptExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_EXCEPT with parameter`() {
+    fun `should support ARRAY_UNION with parameter`() {
         val parameterValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            "ARRAY_EXCEPT($1, `numberArrayField`)",
+            "ARRAY_UNION($1, `numberArrayField`)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayExceptExpression(parameterValue.asParameter(), someNumberArrayField())
+        val underTest = ArrayUnionExpression(parameterValue.asParameter(), someNumberArrayField())
 
         val actual = underTest.toDopeQuery()
 
@@ -44,13 +42,13 @@ class ArrayExceptExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_EXCEPT with parameter as value`() {
+    fun `should support ARRAY_UNION with parameter as value`() {
         val parameterValue = listOf(1, 2, 3)
         val expected = DopeQuery(
-            "ARRAY_EXCEPT(`numberArrayField`, $1)",
+            "ARRAY_UNION(`numberArrayField`, $1)",
             mapOf("$1" to parameterValue),
         )
-        val underTest = ArrayExceptExpression(someNumberArrayField(), parameterValue.asParameter())
+        val underTest = ArrayUnionExpression(someNumberArrayField(), parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -58,14 +56,14 @@ class ArrayExceptExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_EXCEPT with all parameters`() {
+    fun `should support ARRAY_UNION with all parameters`() {
         val parameterValueCollection = listOf(1, 2, 3)
         val parameterValue = listOf(4, 5, 6)
         val expected = DopeQuery(
-            "ARRAY_EXCEPT($1, $2)",
+            "ARRAY_UNION($1, $2)",
             mapOf("$1" to parameterValueCollection, "$2" to parameterValue),
         )
-        val underTest = ArrayExceptExpression(parameterValueCollection.asParameter(), parameterValue.asParameter())
+        val underTest = ArrayUnionExpression(parameterValueCollection.asParameter(), parameterValue.asParameter())
 
         val actual = underTest.toDopeQuery()
 
@@ -73,12 +71,12 @@ class ArrayExceptExpressionTest {
     }
 
     @Test
-    fun `should support ARRAY_EXCEPT extension`() {
+    fun `should support ARRAY_UNION extension`() {
         val array = someNumberArrayField()
         val value = someNumberArrayField()
-        val expected = ArrayExceptExpression(array, value)
+        val expected = ArrayUnionExpression(array, value)
 
-        val actual = arrayExcept(array, value)
+        val actual = arrayUnion(array, value)
 
         assertEquals(expected.toDopeQuery(), actual.toDopeQuery())
     }
